fix(#850): agent_exists INT4 type + canonicalize tmpdir paths in pg review-context tests
- `dispatch_create.rs:438`: sqlx::query_scalar typed as i64 but Postgres
  `SELECT 1` returns INT4 — switch to i32 so `create_dispatch_core`
  agent lookup no longer errors with "mismatched types Rust type i64 ...
  is not compatible with SQL type INT4".
- `dispatch_context.rs` test asserts: tempfile paths on macOS resolve to
  `/private/var/...` after canonicalization while the raw tmpdir handle
  still returns `/var/...`. Compare canonicalized paths on both sides
  instead of raw strings.

+++ src/dispatch/dispatch_create.rs
@@ -435,7 +435,7 @@ async fn create_dispatch_core_internal(
         .map_err(|error| anyhow::anyhow!("Card lookup error: {error}"))?
         .ok_or_else(|| anyhow::anyhow!("Card not found: {kanban_card_id}"))?;
 
-    let agent_exists = sqlx::query_scalar::<_, i64>(
+    let agent_exists = sqlx::query_scalar::<_, i32>(
         "SELECT 1
          FROM agents
          WHERE id = $1
